Moved verify-file-inheritance as a separate tool.
And install the tool as an executable.

This can be useful for anyone who wants to check the status by calling the
tool with a simple system() call. Although in our case, creating the object
on the stack is way easier. That being said, it is useful in a script.